+++ ChangeLog.md
@@ -1,4 +1,5 @@
 2019-11
 
+- reflect feature update when edited fron coords-table
 - include coords-table - preliminary
 - initial commit